refactor: move yandex map functional to MapBlock component
