+++ src/components/MapBlock/MapBlock.scss
@@ -1,0 +1,30 @@
+.map {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+
+  &__button {
+    width: 2.5rem;
+    height: 2.5rem;
+    background-size: 80%;
+    background-repeat: no-repeat;
+    background-position: center;
+
+    &:hover {
+      transform: scale(1.2);
+    }
+  }
+
+  &__close {
+    position: absolute;
+    top: 1rem;
+    right: 1rem;
+    z-index: 3000;
+  }
+
+  &.hidden {
+    display: none;
+  }
+}
